most sold item for merchant seems off
The method makes it seems like this should only return one item (Understand if there are ties can issue a PR in the spec). The `merchant_id` doesn't match the `item_ids` that are located in the spec harness. So just a bit of confusion on how these items are being pulled.

Regards - Allan / Brennan